Extract shared TTS utility components
