+++ sglang_omni/models/higgs_tts/audio_codec.py
@@ -14,7 +14,6 @@ import json
 import os
 from pathlib import Path
 
-import numpy as np
 import torch
 import torch.nn.functional as F
 import torchaudio
@@ -25,8 +24,11 @@ from sglang_omni.models.higgs_tts._vendored.higgs_audio_v2_tokenizer_hf import (
     HiggsAudioV2TokenizerConfig,
     HiggsAudioV2TokenizerModel,
 )
-
-WaveformInput = torch.Tensor | np.ndarray
+from sglang_omni.models.tts_common.codec_batching import (
+    WaveformInput,
+    run_bucketed_batch,
+    to_mono_3d,
+)
 
 # Codec weights live under this prefix inside the TTS checkpoint.
 _CODEC_IN_TTS_CKPT_PREFIX = "tied.embedding.modality_embeddings.0.model."
@@ -40,26 +42,7 @@ _BUNDLED_CODEC_CONFIG_PATH = os.path.join(
 )
 
 
-def _to_mono_3d(waveform: WaveformInput) -> torch.Tensor:
-    """Normalise (Tensor | ndarray) waveform to mono ``[1, 1, L]``."""
-    if isinstance(waveform, np.ndarray):
-        wav = torch.from_numpy(np.ascontiguousarray(waveform))
-    elif isinstance(waveform, torch.Tensor):
-        wav = waveform
-    else:
-        raise TypeError(
-            f"waveform must be Tensor or ndarray, got {type(waveform).__name__}"
-        )
-
-    if wav.ndim == 1:
-        return wav.view(1, 1, -1)
-    if wav.ndim == 2:
-        return wav[:1].unsqueeze(0)
-    if wav.ndim == 3:
-        if wav.shape[1] != 1:
-            raise ValueError(f"audio must be mono, got shape {tuple(wav.shape)}")
-        return wav
-    raise ValueError(f"waveform must be 1-, 2- or 3-D, got {wav.ndim}-D")
+_to_mono_3d = to_mono_3d
 
 
 def _resolve_ckpt_dir(model_path: str | Path) -> str:
@@ -83,7 +66,7 @@ def _load_codec_state_dict(tts_ckpt_dir: str) -> dict[str, torch.Tensor]:
             if full_name.startswith(_CODEC_IN_TTS_CKPT_PREFIX):
                 shards.setdefault(shard, []).append(full_name)
     else:
-        shards = {"model.safetensors": None}  # single-shard layout
+        shards = {"model.safetensors": None}
 
     state: dict[str, torch.Tensor] = {}
     for shard, names in shards.items():
@@ -194,35 +177,16 @@ class HiggsAudioCodec:
         batch_fn,
         error_label: str,
     ) -> list[torch.Tensor]:
-        """Run single_fn on singleton buckets and batch_fn on multi-item buckets."""
-        if not items:
-            return []
-        if len(items) == 1:
-            return [single_fn(items[0])]
-
-        buckets: dict[int, list[int]] = {}
-        for i, item in enumerate(items):
-            buckets.setdefault(bucket_key_fn(item), []).append(i)
-
-        results: list[torch.Tensor | None] = [None] * len(items)
-        for indices in buckets.values():
-            if len(indices) == 1:
-                results[indices[0]] = single_fn(items[indices[0]])
-            else:
-                batch_results = batch_fn([items[i] for i in indices])
-                for idx, result in zip(indices, batch_results):
-                    results[idx] = result
-
-        out: list[torch.Tensor] = []
-        for i, result in enumerate(results):
-            if result is None:
-                raise RuntimeError(f"{error_label} did not produce result for item {i}")
-            out.append(result)
-        return out
+        return run_bucketed_batch(
+            items,
+            bucket_key_fn=bucket_key_fn,
+            single_fn=single_fn,
+            batch_fn=batch_fn,
+            error_label=error_label,
+        )
 
     @torch.no_grad()
     def encode_batch(self, waveforms: list[torch.Tensor]) -> list[torch.Tensor]:
-        # Offline/bulk encoding utility.
         padded: list[torch.Tensor] = []
         for w in waveforms:
             wav = _to_mono_3d(w).to(torch.float32)
