+++ sglang_omni/models/higgs_tts/model_runner.py
@@ -20,7 +20,7 @@ from sglang_omni.model_runner.base import ModelRunner
 from sglang_omni.models.higgs_tts.model import _flat_sampling_attr
 from sglang_omni.models.higgs_tts.sampler import K_MAX
 from sglang_omni.models.higgs_tts.text_tokenizer import AUDIO_PLACEHOLDER_ID
-from sglang_omni.models.higgs_tts.utils import EOC_ID
+from sglang_omni.models.tts_common.codebook_layout import EOC_ID
 from sglang_omni.scheduling.messages import OutgoingMessage
 
 logger = logging.getLogger(__name__)
