+++ sglang_omni/models/higgs_tts/stages.py
@@ -21,12 +21,9 @@ Pipeline shape::
 
 from __future__ import annotations
 
-import base64
 import logging
 import os
 import threading
-from collections import OrderedDict
-from pathlib import Path
 from typing import Any
 
 import torch
@@ -39,17 +36,19 @@ from sglang_omni.models.higgs_tts.payload_types import HiggsTtsState
 from sglang_omni.models.higgs_tts.request_builders import make_higgs_scheduler_adapters
 from sglang_omni.models.higgs_tts.text_tokenizer import HiggsTokenizerAdapter
 from sglang_omni.models.higgs_tts.utils import (
-    apply_delay_pattern,
     get_or_load_codec,
     load_audio_to_24k,
     resolve_checkpoint,
-    to_codes_TN,
     truncate_rope_to_bf16,
 )
 from sglang_omni.models.higgs_tts.vocoder_scheduler import (
     HiggsStreamingVocoderScheduler,
 )
-from sglang_omni.preprocessing.cache_key import hash_bytes, hash_media_item
+from sglang_omni.models.tts_common.codebook_layout import (
+    apply_delay_pattern,
+    to_codes_TN,
+)
+from sglang_omni.models.tts_common.reference_audio import reference_audio_cache_key
 from sglang_omni.proto import StagePayload
 from sglang_omni.scheduling.bootstrap import create_sglang_infrastructure
 from sglang_omni.scheduling.omni_scheduler import OmniScheduler
@@ -71,111 +70,11 @@ _REF_CODE_CACHE_MAX_ITEMS = 256
 _REF_CODE_CACHE_MAX_BYTES = 256 * 1024 * 1024
 _REF_WAVEFORM_CACHE_MAX_ITEMS = 256
 _REF_WAVEFORM_CACHE_MAX_BYTES = 512 * 1024 * 1024
-_REF_PATH_HASH_MEMO_MAX_ITEMS = 1024
-_REF_PATH_HASH_SENTINEL_BYTES = 8192
-_REF_PATH_HASH_MEMO: OrderedDict[str, tuple[str, str]] = OrderedDict()
-_REF_PATH_HASH_MEMO_LOCK = threading.Lock()
 
 # Saturates near c=16 on H100/H200; higher client concurrency only queues.
 DEFAULT_MAX_CONCURRENCY = 16
 
-
-def _reference_path_hash_memo_key(path: Path) -> tuple[str, int] | None:
-    try:
-        if not path.is_file():
-            return None
-        stat_result = path.stat()
-        memo_key = (
-            f"{path.resolve()}:"
-            f"{stat_result.st_size}:"
-            f"{stat_result.st_mtime_ns}:"
-            f"{stat_result.st_ctime_ns}"
-        )
-        return memo_key, int(stat_result.st_size)
-    except OSError:
-        return None
-
-
-def _reference_path_sentinel(path: Path, file_size: int) -> str | None:
-    try:
-        chunk_size = min(_REF_PATH_HASH_SENTINEL_BYTES, file_size)
-        with path.open("rb") as f:
-            chunks = [f.read(chunk_size)]
-            if file_size > _REF_PATH_HASH_SENTINEL_BYTES:
-                middle_offset = max((file_size - chunk_size) // 2, 0)
-                f.seek(middle_offset)
-                chunks.append(f.read(chunk_size))
-            if file_size > 2 * _REF_PATH_HASH_SENTINEL_BYTES:
-                f.seek(max(file_size - chunk_size, 0))
-                chunks.append(f.read(chunk_size))
-        return hash_bytes(b"".join(chunks) + f"|size:{file_size}".encode())
-    except OSError:
-        return None
-
-
-def _get_reference_path_hash(memo_key: str, sentinel: str) -> str | None:
-    with _REF_PATH_HASH_MEMO_LOCK:
-        cached = _REF_PATH_HASH_MEMO.get(memo_key)
-        if cached is None:
-            return None
-        cached_sentinel, digest = cached
-        if cached_sentinel != sentinel:
-            _REF_PATH_HASH_MEMO.pop(memo_key, None)
-            return None
-        _REF_PATH_HASH_MEMO.move_to_end(memo_key)
-        return digest
-
-
-def _put_reference_path_hash(memo_key: str, sentinel: str, digest: str) -> None:
-    with _REF_PATH_HASH_MEMO_LOCK:
-        _REF_PATH_HASH_MEMO[memo_key] = (sentinel, digest)
-        _REF_PATH_HASH_MEMO.move_to_end(memo_key)
-        while len(_REF_PATH_HASH_MEMO) > _REF_PATH_HASH_MEMO_MAX_ITEMS:
-            _REF_PATH_HASH_MEMO.popitem(last=False)
-
-
-def _reference_path_cache_key(path_like: str | Path) -> str | None:
-    # Memoized full-content hash. The stat tuple avoids rereading stable local
-    # refs while still invalidating normal and rapid same-size replacements.
-    path = Path(str(path_like)).expanduser()
-    memo = _reference_path_hash_memo_key(path)
-    if memo is None:
-        return None
-    memo_key, file_size = memo
-    sentinel = _reference_path_sentinel(path, file_size)
-    if sentinel is None:
-        return None
-    digest = _get_reference_path_hash(memo_key, sentinel)
-    if digest is not None:
-        return f"file:{digest}"
-    try:
-        digest = hash_bytes(path.read_bytes())
-    except OSError:
-        return None
-    if _reference_path_hash_memo_key(path) == memo:
-        _put_reference_path_hash(memo_key, sentinel, digest)
-    return f"file:{digest}"
-
-
-def _reference_audio_cache_key(reference_audio: Any) -> str | None:
-    """Stable cache key for a reference-audio input."""
-    if isinstance(reference_audio, (str, Path)):
-        return _reference_path_cache_key(reference_audio)
-    if not isinstance(reference_audio, dict):
-        return None
-    path = reference_audio.get("audio_path") or reference_audio.get("path")
-    if path:
-        return _reference_path_cache_key(path)
-    if "bytes" in reference_audio:
-        data = reference_audio["bytes"]
-        if isinstance(data, str):
-            data = data.encode()
-        return hash_media_item(data)
-    encoded = reference_audio.get("base64") or reference_audio.get("data")
-    if encoded is None:
-        return None
-    raw = base64.b64decode(encoded) if isinstance(encoded, str) else bytes(encoded)
-    return hash_media_item(raw)
+_reference_audio_cache_key = reference_audio_cache_key
 
 
 def create_preprocessing_executor(
