+++ sglang_omni/models/tts_common/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Shared building blocks for TTS model implementations."""
