+++ tests/unit_test/higgs_tts/test_batched_step.py
@@ -18,7 +18,7 @@ from sglang_omni.models.higgs_tts.sampler import (
     batched_step,
     step,
 )
-from sglang_omni.models.higgs_tts.utils import EOC_ID
+from sglang_omni.models.tts_common.codebook_layout import EOC_ID
 
 DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
 # top_k=1 forces greedy (only argmax stays finite after the filter), matching
